Handling glitchy BMC behavior
Sometimes the BMC reports the server's power state is off while
it is actually on. When this situation is caught by Nova compute
manager, it considers the server is shutdown by itself. So the
compute manager proceeds with normal stop API flow, hence the
server is shutdown through IPMI.

In this patch, driver will check the power status again in
power_off method to double confirm its real state. While there's
still possible false alarm situation. Please remind, this is a
workaround.